Enable more flexible serializer structure
This refactors the MapfishPrintProvider class to use a list of associated
serializers when tryong to serialize an array of layers. The elements in the
array are populated with helper methods to register serializers.

All serializer should inherit from the new serializer/Base class, so that
concrete implementations are looking very straightforward.