Change signature of getSerializedLayers
The method is now supposed to be called with a map component as first
argument instead of a layer store. This way we ensure that we have
access to the view resolution for the calls to `serialize()`.

This commit also introduces the possibility to filter the layers to
be serialized by a function. Two new optional arguments (`filterFn`
and `filterScope`) are added to the signature.