Change GeoExt.component.Map initialisation to constructor method
With these changes, the mapcomponent can be used and instanciated
in an universal app, which means it can be used with both toolkits
(classic and modern or ext and touch).

Both toolkits share the constructor method which should be preferred
if possible. The initComponent method will only be available in the
classic toolkit, while the initialize method is only available in
the modern toolkit...