Enable building package w/ sencha cmd on travis
This commit adds several shell scripts, which are used when running
the continuous integration. They download and install all resources
needed for a build using the sencha cmd. As part of the ci, we now
always try to build the GeoExt package.

Additionally the travis ci is configured to cache the contents of
certain directories, which change only every now and then. This
caching speeds up the following builds, where the dependencies for
Sencha or ExtJS have been resolved previously.