Automatically deploy to gh-pages
This commit introduces a shell script that will be run on travis when pushes
to the master happen. It will commit the newest sencha package, API docs and
examples to the gh-pages branch. This will ensure that the homepage is always
up-to-date with the repo.

Pushes to master will end up in a subdirectory master, while named tags will
create a subfolder with the tagname.