Use the `--all` flag when issuing `git add`
This is the fix for the erroneous behaviour noticed in #133 and
circumvented with f0b17ecd1e4deccb39e50d8fe7fd3fcb316338fb.

The behaviour of `git add` has changed between versions of git.
When paths were removed (e.g. with `rm`) then `git add .` would
not add these, but would instead print an error.

The new flag `--all` will also add removed paths to the index,
so that the commit will catch these.

Hat-tip to @dnlkoch for remembering this :heart_eyes:.