Apply correct model to FeatureStore with fields
This ensures that the FeatureStore (GeoExt.data.store.Features) uses the
correct model class in case of beeing instanciated with a 'fields'
configuration. Otherwise an implicit model is generated and the reference
to the OL feature in the record is missing (fixes #171).
This also adds some tests and adapts the featuregrid example to show the
fixed behaviour.