Only publish sencha package when building a tag
Previously we would publish a sencha package for every commit that
happened in the `master`-branch. This commit changes this in a way
that only tagged commits (released versions) will publish a sencha
package.

We still need to define how we actually want to react when we build
for say `v3.0.1`. The current approach would probably make the
originally published `v3.0.0` unusable (because we `rm` probably too
much, and also copy over `catalog.json` metatdata files where we
actually want to merge new `catalog.json` with existing ones).

An upcoming commit will temporary disable the publishing alltogether,
so we cannot break the existing published package.