Introduces ignorePointerRestSelectors
When using the additional config `ignorePointerRestSelectors`, the
pointerrest and pointerrestout events of the `GeoExt.Component.Map`
will be prevented from firing, if the mouse is hovering over an
element that matches at least one of the given css selectors.

This feature comes in handy, e.g. when no events should be fired,
while hovering over tooltips.

Backwards compatibility is maintained by setting `[]` as default
for the config.