Set passThroughFilter to false in WfsFeatureStore
This ensures that the inherited config property is forced to false
within GeoExt.data.store.WfsFeatures. This property only has an effect
if a vector layer is bound, which is not possible for this type of
store.
Furthermore this avoid possible unnecessary requests du to a filter
change in the parent class GeoExt.data.store.Features.